master:

```jl
julia> using ContMechTensors, BenchmarkTools

 julia> v1 = rand(Vec{3}); S1 = rand(SymmetricTensor{4,3}); v2 = rand(Vec{3});

julia> @benchmark dotdot($v1, $S1, $v2)
BenchmarkTools.Trial: 
  memory estimate:  1008.00 bytes
  allocs estimate:  63
  --------------
  minimum time:     271.277 ns (0.00% GC)
  median time:      295.568 ns (0.00% GC)
  mean time:        335.355 ns (7.67% GC)
  maximum time:     3.201 μs (81.83% GC)
  --------------
  samples:          10000
  evals/sample:     303
  time tolerance:   5.00%
  memory tolerance: 1.00%
```

PR:
```jl
julia> using ContMechTensors, BenchmarkTools

 julia> v1 = rand(Vec{3}); S1 = rand(SymmetricTensor{4,3}); v2 = rand(Vec{3});

julia> @benchmark dotdot($v1, $S1, $v2)
BenchmarkTools.Trial: 
  memory estimate:  0.00 bytes
  allocs estimate:  0
  --------------
  minimum time:     36.319 ns (0.00% GC)
  median time:      36.643 ns (0.00% GC)
  mean time:        38.238 ns (0.00% GC)
  maximum time:     96.963 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     990
  time tolerance:   5.00%
  memory tolerance: 1.00%

```